backend/api/requirements.txt
asgiref==3.5.2
astroid==2.12.10
dill==0.3.5.1
Django==4.1.1
djangorestframework==3.14.0
dodgy==0.2.1
flake8==4.0.1
flake8-polyfill==1.0.2
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
mypy==0.971
mypy-extensions==0.4.3
pep8-naming==0.10.0
platformdirs==2.5.2
pycodestyle==2.8.0
pydocstyle==6.1.1
pyflakes==2.4.0
pylint==2.15.3
pylint-celery==0.3
pylint-django==2.5.3
pylint-flask==0.6
pylint-plugin-utils==0.7
pytz==2022.2.1
PyYAML==6.0
requirements-detector==0.7
setoptconf-tmp==0.3.1
snowballstemmer==2.2.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
typing_extensions==4.3.0
wrapt==1.14.1
